store: type-assert int64 for assert tests

+++ store/store_test.go
@@ -62,7 +62,7 @@ func TestMinExpireTime(t *testing.T) {
 	assert.Equal(t, e.EtcdIndex, eidx, "")
 	assert.Equal(t, e.Action, "get", "")
 	assert.Equal(t, e.Node.Key, "/foo", "")
-	assert.Equal(t, e.Node.TTL, 0)
+	assert.Equal(t, e.Node.TTL, int64(0))
 }
 
 // Ensure that the store can recursively retrieve a directory listing.
@@ -107,7 +107,7 @@ func TestStoreGetDirectory(t *testing.T) {
 		case "/foo/baz/ttl":
 			assert.Equal(t, *node.Value, "Y", "")
 			assert.Equal(t, node.Dir, false, "")
-			assert.Equal(t, node.TTL, 3, "")
+			assert.Equal(t, node.TTL, int64(3), "")
 		default:
 			t.Errorf("key = %s, not matched", node.Key)
 		}
@@ -162,7 +162,7 @@ func TestSet(t *testing.T) {
 	assert.Equal(t, *e.Node.Value, "", "")
 	assert.Nil(t, e.Node.Nodes, "")
 	assert.Nil(t, e.Node.Expiration, "")
-	assert.Equal(t, e.Node.TTL, 0, "")
+	assert.Equal(t, e.Node.TTL, int64(0), "")
 	assert.Equal(t, e.Node.ModifiedIndex, uint64(1), "")
 
 	// Set /foo="bar"
@@ -176,7 +176,7 @@ func TestSet(t *testing.T) {
 	assert.Equal(t, *e.Node.Value, "bar", "")
 	assert.Nil(t, e.Node.Nodes, "")
 	assert.Nil(t, e.Node.Expiration, "")
-	assert.Equal(t, e.Node.TTL, 0, "")
+	assert.Equal(t, e.Node.TTL, int64(0), "")
 	assert.Equal(t, e.Node.ModifiedIndex, uint64(2), "")
 	// check prevNode
 	assert.NotNil(t, e.PrevNode, "")
@@ -194,7 +194,7 @@ func TestSet(t *testing.T) {
 	assert.Equal(t, *e.Node.Value, "baz", "")
 	assert.Nil(t, e.Node.Nodes, "")
 	assert.Nil(t, e.Node.Expiration, "")
-	assert.Equal(t, e.Node.TTL, 0, "")
+	assert.Equal(t, e.Node.TTL, int64(0), "")
 	assert.Equal(t, e.Node.ModifiedIndex, uint64(3), "")
 	// check prevNode
 	assert.NotNil(t, e.PrevNode, "")
@@ -213,7 +213,7 @@ func TestSet(t *testing.T) {
 	assert.Nil(t, e.Node.Value)
 	assert.Nil(t, e.Node.Nodes, "")
 	assert.Nil(t, e.Node.Expiration, "")
-	assert.Equal(t, e.Node.TTL, 0, "")
+	assert.Equal(t, e.Node.TTL, int64(0), "")
 	assert.Equal(t, e.Node.ModifiedIndex, uint64(4), "")
 }
 
@@ -231,7 +231,7 @@ func TestStoreCreateValue(t *testing.T) {
 	assert.Equal(t, *e.Node.Value, "bar", "")
 	assert.Nil(t, e.Node.Nodes, "")
 	assert.Nil(t, e.Node.Expiration, "")
-	assert.Equal(t, e.Node.TTL, 0, "")
+	assert.Equal(t, e.Node.TTL, int64(0), "")
 	assert.Equal(t, e.Node.ModifiedIndex, uint64(1), "")
 
 	// Create /empty=""
@@ -245,7 +245,7 @@ func TestStoreCreateValue(t *testing.T) {
 	assert.Equal(t, *e.Node.Value, "", "")
 	assert.Nil(t, e.Node.Nodes, "")
 	assert.Nil(t, e.Node.Expiration, "")
-	assert.Equal(t, e.Node.TTL, 0, "")
+	assert.Equal(t, e.Node.TTL, int64(0), "")
 	assert.Equal(t, e.Node.ModifiedIndex, uint64(2), "")
 
 }
@@ -292,12 +292,12 @@ func TestStoreUpdateValue(t *testing.T) {
 	assert.Equal(t, e.Node.Key, "/foo", "")
 	assert.False(t, e.Node.Dir, "")
 	assert.Equal(t, *e.Node.Value, "baz", "")
-	assert.Equal(t, e.Node.TTL, 0, "")
+	assert.Equal(t, e.Node.TTL, int64(0), "")
 	assert.Equal(t, e.Node.ModifiedIndex, uint64(2), "")
 	// check prevNode
 	assert.Equal(t, e.PrevNode.Key, "/foo", "")
 	assert.Equal(t, *e.PrevNode.Value, "bar", "")
-	assert.Equal(t, e.PrevNode.TTL, 0, "")
+	assert.Equal(t, e.PrevNode.TTL, int64(0), "")
 	assert.Equal(t, e.PrevNode.ModifiedIndex, uint64(1), "")
 
 	e, _ = s.Get("/foo", false, false)
@@ -313,12 +313,12 @@ func TestStoreUpdateValue(t *testing.T) {
 	assert.Equal(t, e.Node.Key, "/foo", "")
 	assert.False(t, e.Node.Dir, "")
 	assert.Equal(t, *e.Node.Value, "", "")
-	assert.Equal(t, e.Node.TTL, 0, "")
+	assert.Equal(t, e.Node.TTL, int64(0), "")
 	assert.Equal(t, e.Node.ModifiedIndex, uint64(3), "")
 	// check prevNode
 	assert.Equal(t, e.PrevNode.Key, "/foo", "")
 	assert.Equal(t, *e.PrevNode.Value, "baz", "")
-	assert.Equal(t, e.PrevNode.TTL, 0, "")
+	assert.Equal(t, e.PrevNode.TTL, int64(0), "")
 	assert.Equal(t, e.PrevNode.ModifiedIndex, uint64(2), "")
 
 	e, _ = s.Get("/foo", false, false)
